MET-543 Fix, remove logging the exception when custom truststore fails.
To not have huge amount of logs for repeating requests to resources that do not use the custom truststore certificates, the exception doesn't have to be logged. If that happens the exception is caught and another try to use the default trustore will occur which in turn will fail if there are no valid certificates.