MET-3306 Remove unused code
Removed the wikidata code that was unused internal and/or externally. The code present there was used once in the past and we have updated it multiple times to conform our changes in enrichment but it was never used. Once we decide about enriching/dereferencing organizations the implementation won't be relevant with the current code for wikidata.